Drupal 7: only delete files when exists.

+++ src/Behat/Cores/Drupal7.php
@@ -158,7 +158,9 @@ class Drupal7 extends OriginalDrupal7 implements CoreInterface {
    */
   public function fileDelete($fid) {
     $file = file_load($fid);
-    file_delete($file);
+    if ($file) {
+      file_delete($file);
+    }
   }
 
 }
